Changed the variable c1 in returnData to a double in boxvalues.h as need the decimal

+++ boxValues.h
@@ -6,7 +6,7 @@ namespace boxValues{
 	struct returnData {
 		double v1;
 		short r1, r2;
-		short c1;
+		double c1;
 		double f;
 		byte error;
 	};
